Split orchestration generation module into focused submodules
The 848-line generation.py module mixed enums, exceptions, validators,
DTOs, protocols, the show-notes tool executor, and the planner /
orchestrator drivers in one file. Split into four cohesive modules so
each stays well below 400 lines while keeping
episodic/orchestration/__init__.py unchanged:

- _types.py: enums, exception hierarchy, and the shared _log_event
  helper.
- _dto.py: validators, frozen dataclasses, protocol ports, and the
  build_generation_result aggregator. Annotations are deferred via
  PEP 563 so ShowNotesResult can stay in a TYPE_CHECKING import.
- _show_notes_executor.py: ShowNotesToolExecutor and its private
  helpers.
- generation.py: re-exports the public surface and hosts the two
  remaining drivers, StructuredGenerationPlanner and
  StructuredPlanningOrchestrator.
